N21/idm (#820)
* min-height for content - use available space

* Transfer school information to server for ldap authentication

* Transfer school information to server for ldap authentication (#818)

* firstlogin modularization #1

* firstLogin modularization #2

* firstLogin modularization #2.1 (bugfix)

* firstLogin modularization #3

added existing users and some routing

* bugfixing

* added redirects & naming changes

* u14 bugfix

* geb14 bugfix

* minor layout change

* new TODOS & fixes after meeting with Konstantin

* wip

* SC-206 ~ Add basic LibreOffice Online Client (#731)

* add basic libre office client on file page

* for now set other wopirestUrl

* add accesToken

* change way of opening doc-files

* lool just for doc-files

* add open office doc

* create docx, pptx, xlsx

* remove double id

* remove double slash

* shareable view/download link

* course file edit restrictions

* restrict course file creation

* add studentCanEdit for existing office files

* finish icon

* minor changes

* jans request

* Update files.js

* add ods (spreadsheet) and ods (presentation) handling

* Update gulpfile.js (#827)

* more modularization (including some cleanup)

* removed console.log

* Hide login system select if only one system exists

* fix cloudname not showing

* redirect to new firstLogin

* move submit logic to server

* Disable editing account details when logged in via SSO

* added prefix option

* added birthday routing

* fix redirect, fix logic to show parentConsent page

* added missing import

* fix endless redirects

* fix firstLogin in case of sso
