+++ .eslintrc.json
@@ -13,8 +13,6 @@
     "it": true,
     "populateModalForm": true,
     "showAJAXError": true,
-    "getQueryParameterByName": true,
-    "updateQueryStringParameter": true,
     "populateModal": true,
     "printPart": true
   },
