+++ static/styles/administration/dataprivacy.scss
@@ -1,5 +1,6 @@
 @import 'progress';
 @import '../lib/colors';
+@import '../lib/datetimepicker/jquery.datetimepicker.min.css';
 
 .panels{
     /* insert custom css for the forms here */
