+++ static/styles/administration/progress.scss
@@ -91,7 +91,7 @@
   background: #c0c0c0;
   display: inline-block;
   height: 3px;
-  transform: translateY(-2.5em);
+  transform: translateY(-3em);
   transition: 0.3s;
   width: 0;
 }
