+++ static/styles/authentication/home.scss
@@ -71,7 +71,7 @@ body {
         }
     }
     main{
-        min-height: calc(100vh - 228px );
+        min-height: calc(100vh - 228px);
         .illustration{
             min-width: 150px;
             width: 25vw;
