+++ static/styles/lib/base.scss
@@ -137,19 +137,7 @@ footer {
 		margin-left:-11px;
 	}
 	&.small-footer {
-		width: 100%;
-		margin-top: 40px;
-		position: inherit;
-		@media screen and (max-height: 700px) {
-			top: 510px;
-		}
-		bottom: 0;
-		right: 0;
-		left: 0;
-		clear: both;
-		padding: 30px 0px;
-		font-size: 0.9em;
-		margin-bottom: 5px;
+
 		@include media-breakpoint-down(xs) {
 			background: $colorBeige;
