+++ static/styles/lib/loggedin.scss
@@ -286,12 +286,6 @@ aside.nav-sidebar {
     }
 }
 
-.content-min-height{
-    min-height: calc(100vh - 193px);
-    @include media-breakpoint-down(xs) {
-        min-height: calc(100vh - 244px);
-    }
-}
 header {
     width:100%;
     float:left;
@@ -301,21 +295,36 @@ header {
     padding: 20px !important;
 }
 
+.content-min-height {
+    flex:1;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+
+    #main-content {
+        flex:1;
+        position: relative;
+        padding: 0;
+        margin-left: 15px;
+        margin-right: 15px;
+    }
+}
+
 .content-wrapper {
     vertical-align: top;
-    width: calc(100% - 240px);
-    margin-left: 240px;
-    overflow: auto;
+    width: 100%;
     padding: 0;
+    padding-left: 240px;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
 
     @include media-breakpoint-down(md) {
-        width: calc(100% - 60px);
-        margin-left: 60px;
+        padding-left: 60px;
     }
 
     @include media-breakpoint-down(xs) {
-        width: 100%;
-        margin-left: 0px;
+        padding-left: 0px;
     }
     .school-data {
         padding-top: 11px;
@@ -371,10 +380,7 @@ body.fullscreen{
     }
     .content-wrapper{
         width: 100vw;
-        margin-left: 0;
-        .content-min-height{
-            min-height: 100vh;
-        }
+        padding-left: 0;
     }
     .btn-fullscreen{
         position: fixed;
@@ -394,12 +400,7 @@ body.fullscreen{
 
 footer {
     width: 100%;
-    position: inherit;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    clear: both;
-    padding: 30px 0px;
+    padding: 20px 0px;
 
     @include media-breakpoint-down(xs) {
         background: $colorBeige;
