+++ views/dashboard/dashboard.hbs
@@ -10,7 +10,7 @@
     {{/content}}
 
     {{#content "page"}}
-        <section class="section-timetable" data-step="7" data-intro="Hier findest du eine Übersicht über deinen Stundenplan und anstehende Termine.">
+        <section class="container-fluid section-timetable" data-step="7" data-intro="Hier findest du eine Übersicht über deinen Stundenplan und anstehende Termine.">
             <div class="row">
                 <div class="col-sm-12">
                     <h5>Stundenplan</h5>
@@ -18,7 +18,7 @@
                 </div>
             </div>
 
-            <div class="container-fluid timetable">
+            <div class="timetable">
                 <div class="row lesson-cards">
                     <div>
                         {{#each events}}
