+++ views/firstLogin/firstLoginExistingEmployee.hbs
@@ -1,36 +1,15 @@
 {{#extend "lib/simple"}}
-  {{#content "scripts" mode="append"}}
-      <script src="/scripts/dataprivacy/dataprivacy.js" type="text/javascript" defer></script>
-      <script src="/scripts/jquery/jquery.datetimepicker.full.min.js" type="text/javascript" defer></script>
-  {{/content}}
-  {{#content "styles" mode="append"}}
-      <link rel="stylesheet" href="/styles/administration/dataprivacy.css"/>
-      <link rel="stylesheet" href="/styles/lib/datetimepicker/jquery.datetimepicker.min.css"/>
- {{/content}}
+  {{> "firstLogin/partials/imports"}}
   {{#content "page"}}
     {{#embed "lib/forms/paginatedForm" sections=3 action="/firstLogin/submit" method="POST"}}
       {{#content "panels"}}
-      <section data-panel="section-1">
-        <h4>Willkommen zurück in der {{theme.title}}, {{currentUser.firstName}} {{currentUser.lastName}}</h4>
-       <p>Mit dem Schuljahr 2018/2019 haben wir die Registrierung in der {{theme.title}} von Papier auf Digital umgestellt. Bitte bestätige im nächsten Schritt die Einwilligungserklärungen, welche du schon auf Papier bestätigt hast, erneut, damit du die {{theme.short_title}} weiter nutzen kannst.</p>
-      </section>
 
+        {{> "firstLogin/sections/welcome_existing" section-nr=1}}
 
-      <section data-panel="section-2" class="submit-page">
-        <h4>Einwilligungserklärung</h4>
-        <p>Du bist mindestens 14 Jahre alt. Deine Erziehungsberechtigten (oder du selbst, wenn du schon über 18 bist) haben der Nutzung bereits auf dem Papierweg zugestimmt. Bitte bestätige jetzt folgende Einwilligungserklärungen, damit du die {{theme.title}} weiter nutzen kannst.</p>
+        {{> "firstLogin/sections/consent" section-nr=2 submit-page=true}}
 
-          <p>Ich, <b>{{currentUser.firstName}} {{currentUser.lastName}}</b>, erkläre Folgendes:</p>
-          {{#embed "registration/forms/einwilligungserklaerung"}}{{/embed}}
-      </section>
+        {{> "firstLogin/sections/thanks" section-nr=3}}
 
-
-      <section data-panel="section-3">
-        <h4>Danke!</h4>
-        <p>Du kannst jetzt mit der Benutzung der {{theme.short_title}} fortfahren.<br>
-            <a class="btn btn-primary mt-2" href="/login">Weiter zur {{theme.short_title}}</a>
-        </p>
-      </section>
       {{/content}}
     {{/embed}}
   {{/content}}
