--- views/firstLogin/firstLoginUE18.hbs
@@ -1,64 +1,0 @@
-{{#extend "lib/simple"}}
-  {{#content "scripts" mode="append"}}
-      <script src="/scripts/dataprivacy/dataprivacy.js" type="text/javascript" defer></script>
-      <script src="/scripts/pwrecovery.js" type="text/javascript" defer></script>
-      <script src="/scripts/jquery/jquery.datetimepicker.full.min.js" type="text/javascript" defer></script>
-  {{/content}}
-  {{#content "styles" mode="append"}}
-      <link rel="stylesheet" href="/styles/administration/dataprivacy.css"/>
-      <link rel="stylesheet" href="/styles/lib/datetimepicker/jquery.datetimepicker.min.css"/>
- {{/content}}
-  {{#content "page"}}
-    {{#embed "lib/forms/paginatedForm" sections=3 action="/firstLogin/submit" method="post"}}
-      {{#content "panels"}}
-
-      <section data-panel="section-1">
-        <h4>Herzlich Willkommen in der {{theme.title}}, {{currentUser.firstName}} {{currentUser.lastName}}</h4>
-        <p>Wir freuen uns, dass du die {{theme.short_title}} nutzen möchtest.</p>
-        <p>Bitte ändere im nächsten Schritt dein Startpasswort, damit du mit der Benutzung der {{theme.short_title}} starten kannst.</p>
-        <p class="mt-2 explanation nomargin">* Sprachlicher Hinweis: Für bessere Lesbarkeit verwenden wir in der {{theme.short_title}} die männliche Form (z.B. „Schüler“, „Lehrer“). Es sind stets alle Personen unabhängig vom Geschlecht gemeint. Die {{theme.short_title}} ist für alle Menschen da, unabhängig von Herkunft, Einschränkungen, Geschlecht und sexueller Orientierung. 🌈🎉🦄 </p>
-     {{#if sso}}<input type="hidden" name="sso" value="{{sso}}" />{{/if}}
-	 </section>
-
-      <section data-panel="section-2" class="submit-page">
-        <h4>Du hast es fast geschafft...</h4>
-		{{#if sso}}
-		 <p>Nach klicken des <u>Absenden</u>-Buttons kannst du dich wie gewohnt über das System deiner Schule (Moodle, iServ, ItsLearning etc.) in der {{theme.title}} einloggen. </p>
-	    {{else}}
-        <p>Bitte ändere das Startpasswort. Notiere dir das neue Passwort gleich und bewahre es gut auf!</p>
-        <div class="row">
-          <div class="form-group col-md-6 col-sm-12">
-            <label>Altes Passwort: *</label>
-            <input class="form-control" id="password_verification" name="password_verification" type="password" placeholder="*******" required/>
-          </div>
-        </div>
-
-        <p>Mindestens 8 Zeichen mit Groß- und Kleinschreibung und mindestens einer Zahl und Sonderzeichen. <br/>
-          Erlaubte Sonderzeichen sind: <code>!§$%&/()=?\;:,.#+*~</code></p>
-        <div class="row">
-          <div class="form-group col-md-6 col-sm-12">
-            <label>Neues Passwort: *</label>
-            <input class="form-control" id="password" name="password-1" type="password" placeholder="Neues Passwort" minlength="8" pattern="^(?=.*[A-Z])(?=.*[0-9])(?=.*[a-z])(?=.*[\-_!<>§$%&\/()=?\\;:,.#+*~']).{8,255}$" title="Mindestens 8 Zeichen mit Groß- und Kleinschreibung und mindestens einer Zahl und Sonderzeichen." required/>
-          </div>
-          <div class="form-group col-md-6 col-sm-12">
-            <label>Wiederhole dein neues Passwort: *</label>
-            <input class="form-control" id="password_control" name="password-2" type="password" placeholder="Neues Passwort wiederholen" required/>
-          </div>
-        </div>
-
-          <!--<label>
-              <input name="sendCredentials" type="checkbox" value="sendCredentials"/>
-              Login-Daten per Mail an mich senden
-          </label>-->
-		{{/if}}
-      </section>
-
-      <section data-panel="section-3">
-        <h4>Fertig!</h4>
-        <p>Du kannst jetzt mit der Benutzung der {{theme.title}} starten.</p>
-        <a class="btn btn-primary mt-2" href="/login">{{theme.short_title}} erkunden</a>
-      </section>
-      {{/content}}
-    {{/embed}}
-  {{/content}}
-{{/extend}}
