+++ views/firstLogin/partials/imports.hbs
@@ -1,0 +1,8 @@
+{{#content "scripts" mode="append"}}
+  <script src="/scripts/dataprivacy/dataprivacy.js" type="text/javascript" defer></script>
+  <script src="/scripts/confetti/confetti.js" type="text/javascript" defer></script>
+{{/content}}
+{{#content "styles" mode="append"}}
+  <link rel="stylesheet" href="/styles/administration/dataprivacy.css"/>
+  <link rel="stylesheet" href="/styles/lib/confetti/confetti.css"/>
+{{/content}}
