+++ views/firstLogin/partials/welcome_existing.hbs
@@ -1,0 +1,2 @@
+<p>Herzlich Willkommen zurück in der {{@root.theme.short_title}}, {{@root.currentUser.firstName}} {{@root.currentUser.lastName}},</p>
+<p>mit dem Schuljahr 2018/2019 haben wir die Registrierung in der {{@root.theme.short_title}} von Papier auf Digital umgestellt. Das macht es notwendig, dass wir dein Einverständnis zur Nutzung der {{@root.theme.short_title}} erneut elektronisch erfassen.</p>
