+++ views/firstLogin/sections/email.hbs
@@ -1,0 +1,11 @@
+<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
+  <h4>Wie können wir dich erreichen?</h4>
+  <p>Bitte überprüfe deine E-Mail-Adresse:</p>
+  <div class="row">
+    <div class="form-group col-md-6 col-sm-12">
+        <label>E-Mail:</label>
+        <input class="form-control linked" type="email" name="student-email" placeholder="schueler@schul-cloud.org" value="{{@root.currentUser.email}}" readonly required>
+    </div>
+  </div>
+  <p>Sollte die E-Mail-Adresse nicht die richtige sein, wende dich bitte an deinen {{@root.theme.short_title}}-Administrator.</p>
+</section>
