+++ views/firstLogin/sections/parent_intro.hbs
@@ -1,0 +1,6 @@
+<section data-panel="section-{{section-nr}}" class="parent_intro" data-next-label="als Erziehungsberechtigter fortfahren">
+  <h4>Einverständniserklärung eines Erziehungsberechtigten benötigt</h4>
+  <p>Da du unter 18 Jahre alt bist, wird die Einverständniserklärung eines Elternteils benötigt, damit du die {{root.theme.short_title}} nutzen kannst. 
+    Die nächsten Schritte müssen von einem Erziehungsberechtigten durchgeführt werden.</p>
+  <a href="/logout" class="btn btn-secondary">Beenden und später fortfahren</a>
+</section>
