+++ views/firstLogin/sections/password.hbs
@@ -1,0 +1,27 @@
+<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
+  <h4>Du hast es fast geschafft...</h4>
+  {{#if @root.sso}}
+    <p>Nach klicken des <u>Absenden</u>-Buttons kannst du dich wie gewohnt über das System deiner Schule (Moodle, iServ, ItsLearning etc.) in der {{@root.theme.title}} einloggen. </p>
+  {{else}}
+    <p>Bitte ändere das Startpasswort. Notiere dir das neue Passwort gleich und bewahre es gut auf!</p>
+    <p>
+      Mindestens 8 Zeichen mit Groß- und Kleinschreibung, davon jeweils mindestens eine Zahl und ein Sonderzeichen. <br/>
+      Erlaubte Sonderzeichen sind: <code>!§$%&/()=?\;:,.#+*~</code>
+    </p>
+    <div class="row">
+      <div class="form-group col-md-6 col-sm-12">
+        <label>Neues Passwort: *</label>
+        <input class="form-control" id="password" name="password-1" type="password" placeholder="Neues Passwort" minlength="8" pattern="^(?=.*[A-Z])(?=.*[0-9])(?=.*[a-z])(?=.*[\-_!<>§$%&\/()=?\\;:,.#+*~']).{8,255}$" title="Mindestens 8 Zeichen mit Groß- und Kleinschreibung und mindestens einer Zahl und Sonderzeichen." required/>
+      </div>
+      <div class="form-group col-md-6 col-sm-12">
+        <label>Wiederhole dein neues Passwort: *</label>
+        <input class="form-control" id="password_control" name="password-2" type="password" placeholder="Neues Passwort wiederholen" required/>
+      </div>
+    </div>
+
+    <!--<label>
+        <input name="sendCredentials" type="checkbox" value="sendCredentials"/>
+        Login-Daten per Mail an mich senden
+    </label>-->
+  {{/if}}
+</section>
