+++ views/firstLogin/sections/thanks.hbs
@@ -1,0 +1,6 @@
+<section data-panel="section-{{section-nr}}">
+    <h4>Danke!</h4>
+    <p>Du kannst jetzt mit der Benutzung der {{@root.theme.short_title}} fortfahren.<br>
+        <a class="btn btn-primary mt-2" href="/">{{@root.theme.short_title}} erkunden</a>
+    </p>
+</section>
