+++ views/firstLogin/sections/welcome.hbs
@@ -1,0 +1,7 @@
+<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
+  <h4>Herzlich Willkommen in der {{@root.theme.short_title}}, {{@root.currentUser.firstName}} {{@root.currentUser.lastName}}</h4>
+  <p>Wir freuen uns, dass du die {{@root.theme.short_title}} nutzen möchtest.</p>
+  <p>Bitte gleiche kurz ein paar Daten ab, lege ein neues Passwort fest und dann kann es auch schon losgehen. 😊</p>
+  {{> "firstLogin/partials/language-hint"}}
+  {{#if @root.sso}}<input type="hidden" name="sso" value="{{@root.sso}}" />{{/if}}
+</section>
