+++ views/firstLogin/sections/welcome_14-17.hbs
@@ -1,0 +1,9 @@
+<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
+  <h4>Herzlich Willkommen in der {{@root.theme.short_title}}, {{@root.currentUser.firstName}} {{@root.currentUser.lastName}} </h4>
+  <p>
+      Wir freuen uns, dass du die {{@root.theme.title}} nutzen möchtest und mithilfst, gemeinsam den nächsten Schritt zur digitalen Unterstützung des Unterrichts zu gehen – Open Source, gemeinnützig und datenschutzkonform.
+      Bitte überprüfe in den kommenden Schritten die für dich angegebene E-Mail-Adresse und stimme den Nutzungsbedingungen zu. Ein Erziehungsberechtigter hat der Nutzung bereits zugestimmt, jetzt bist du gefragt. Ändere anschließend dein Startpasswort, damit du die Vorteile der {{@root.theme.short_title}} nutzen kannst.
+  </p>
+  {{> "firstLogin/partials/language-hint"}}
+  {{#if @root.sso}}<input type="hidden" name="sso" value="{{@root.sso}}" />{{/if}}
+</section>
