+++ views/firstLogin/sections/welcome_existing.hbs
@@ -1,0 +1,4 @@
+<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
+    {{> "firstLogin/partials/welcome_existing"}}
+    {{> "firstLogin/partials/language-hint"}}
+</section>
