+++ views/firstLogin/sections/welcome_existing_geb14.hbs
@@ -1,0 +1,4 @@
+<section data-panel="section-{{section-nr}}" class="confetti-wrapper{{#if submit-page}} submit-page{{/if}}">
+  <h4>Herzlichen Glückwunsch, {{@root.currentUser.firstName}} {{@root.currentUser.lastName}}! 🎉🎈</h4>
+  <p>Du bist jetzt 14 Jahre alt. Bitte erkläre im folgenden Schritt dein Einverständnis zur Nutzung der {{@root.theme.title}}. Ab 14 Jahren ist dies zusätzlich zum bereits erklärten Einverständnis deiner Eltern nötig, damit du die {{@root.theme.short_title}} weiter nutzen kannst.</p>
+</section>
