+++ views/registration/forms/einwilligungserklaerung.hbs
@@ -1,23 +1,23 @@
 <label>
-    <input name="privacyConsent" type="checkbox" value="true" required/>
+    <input name="{{prefix}}privacyConsent" type="checkbox" value="true" required/>
     <span>Ich habe die <a href="https://schul-cloud-hpi.obs.eu-de.otc.t-systems.com:443/Dokumente/Einwilligung/1-Datenschutzerklaerung.pdf" target="_blank">Datenschutzerklärung</a> gelesen und willige in die Verarbeitung der Nutzerdaten ein.</span>
 </label>
-<p class="explanation">Um die Nutzung der {{theme.short_title}} zu ermöglichen, ist die Verarbeitung von personenbezogenen Daten der Nutzer erforderlich. Dies erfolgt ausschließlich entsprechend hoher gesetzlicher Datenschutz-Anforderungen. Das HPI verarbeitet nur die Daten, deren Erhebung unbedingt erforderlich ist und sorgt dafür, dass sie nicht in unbefugte Hände geraten. </p>
+<p class="explanation">Um die Nutzung der {{@root.theme.short_title}} zu ermöglichen, ist die Verarbeitung von personenbezogenen Daten der Nutzer erforderlich. Dies erfolgt ausschließlich entsprechend hoher gesetzlicher Datenschutz-Anforderungen. Das HPI verarbeitet nur die Daten, deren Erhebung unbedingt erforderlich ist und sorgt dafür, dass sie nicht in unbefugte Hände geraten. </p>
 
 <label>
-    <input name="researchConsent" type="checkbox" value="true" required/>
+    <input name="{{prefix}}researchConsent" type="checkbox" value="true" required/>
     <span>Ich habe die <a href="https://schul-cloud-hpi.obs.eu-de.otc.t-systems.com:443/Dokumente/Einwilligung/2-Information-zum-Forschungsvorhaben.pdf" target="_blank">Informationen zum Forschungsvorhaben</a> gelesen und willige in die wissenschaftliche Auswertung ein.</span>
 </label>
-<p class="explanation">Die {{theme.title}} wird von einem wissenschaftlichen Forschungsprojekt begleitet, das der Optimierung des Lernerlebnisses in der {{theme.short_title}} dient. Im Rahmen dessen werden Nutzungsdaten anonymisiert ausgewertet.</p>
+<p class="explanation">Die {{@root.theme.title}} wird von einem wissenschaftlichen Forschungsprojekt begleitet, das der Optimierung des Lernerlebnisses in der {{@root.theme.short_title}} dient. Im Rahmen dessen werden Nutzungsdaten anonymisiert ausgewertet.</p>
 
 <label>
-    <input name="thirdPartyConsent" type="checkbox" value="true" required/>
+    <input name="{{prefix}}thirdPartyConsent" type="checkbox" value="true" required/>
     <span> Ich habe die <a href="https://schul-cloud-hpi.obs.eu-de.otc.t-systems.com:443/Dokumente/Einwilligung/3-Informationen-zu-Inhalte-Anbietern.pdf" target="_blank">Informationen zu Inhalte-Anbietern</a> gelesen und willige in die Weitergabe der Pseudonyme an Inhalte-Anbieter ein</span>
 </label>
-<p class="explanation">Nutzernamen werden nicht an integrierte Drittanbieter weitergegeben. Sollte eine Wiedererkennung nötig sein, z.B. für individuelle Aufgabenstellungen, geschieht dies zum Schutz der Nutzeridentität auf Basis eines Pseudonyms. Die Nutzerdaten bleiben sicher im Kernbereich der {{theme.short_title}}.</p>
+<p class="explanation">Nutzernamen werden nicht an integrierte Drittanbieter weitergegeben. Sollte eine Wiedererkennung nötig sein, z.B. für individuelle Aufgabenstellungen, geschieht dies zum Schutz der Nutzeridentität auf Basis eines Pseudonyms. Die Nutzerdaten bleiben sicher im Kernbereich der {{@root.theme.short_title}}.</p>
 
 <label>
-    <input name="termsOfUseConsent" type="checkbox" value="true" required/>
+    <input name="{{prefix}}termsOfUseConsent" type="checkbox" value="true" required/>
     <span>Ich habe die <a href="https://schul-cloud-hpi.obs.eu-de.otc.t-systems.com:443/Dokumente/Einwilligung/4-Nutzungsbedingungen.pdf" target="_blank">Nutzungsbedingungen</a> gelesen und stimme ihnen zu.</span>
 </label>
-<p class="explanation">Die Nutzung der {{theme.short_title}} ist freiwillig. Die erteilte Einwilligung kann daher jederzeit ohne Nachteile für den Nutzer widerrufen werden. Eine kurze Mail mit Angabe des Vor- und Nachnamens, der Klasse und der Schule an folgende Adresse genügt: <a href="mailto:widerruf@schul-cloud.org?subject=Widerruf der Einwilligung&body=Liebes {{theme.short_title}}-Team,%0D%0Ahiermit widerrufe ich die von mir erteilte EInwilligung zur Verarbeitung meiner personenbezogenen Daten.%0D%0A%0D%0AVor- und Nachname:%0D%0AKlasse:%0D%0ASchule:%0D%0A">widerruf@schul-cloud.org</a></p>
+<p class="explanation">Die Nutzung der {{@root.theme.short_title}} ist freiwillig. Die erteilte Einwilligung kann daher jederzeit ohne Nachteile für den Nutzer widerrufen werden. Eine kurze Mail mit Angabe des Vor- und Nachnamens, der Klasse und der Schule an folgende Adresse genügt: <a href="mailto:widerruf@schul-cloud.org?subject=Widerruf der Einwilligung&body=Liebes {{@root.theme.short_title}}-Team,%0D%0Ahiermit widerrufe ich die von mir erteilte EInwilligung zur Verarbeitung meiner personenbezogenen Daten.%0D%0A%0D%0AVor- und Nachname:%0D%0AKlasse:%0D%0ASchule:%0D%0A">widerruf@schul-cloud.org</a></p>
