+++ views/registration/registration-employee.hbs
@@ -32,7 +32,7 @@
 
 
       <section data-panel="section-2">
-        <h4>Ihre Daten</h4>
+        <h4>Deine Daten</h4>
         <p>Die {{theme.short_title}} bewahrt Nutzerdaten sicher auf und gibt sie nicht an Dritte weiter. Die Verarbeitung der Daten erfolgt entsprechend der hohen gesetzlichen Datenschutz-Anforderungen.</p>
 
         <p>Für die Nutzung der {{theme.short_title}} benötigen wir folgende Angaben:</p>
